requirements.txt
Django==1.8
django-authtools==1.2.0
